+++ packages/gatsby/src/db/__tests__/nodes.js
@@ -3,10 +3,14 @@ const { getNode, getNodes } = require(`../nodes`)
 const { store } = require(`../../redux`)
 require(`./fixtures/ensure-loki`)()
 
+const report = require(`gatsby-cli/lib/reporter`)
+jest.mock(`gatsby-cli/lib/reporter`)
+
 describe(`nodes db tests`, () => {
   beforeEach(() => {
     store.dispatch({ type: `DELETE_CACHE` })
   })
+
   it(`deletes previously transformed children nodes when the parent node is updated`, () => {
     store.dispatch(
       actions.createNode(
@@ -196,6 +200,7 @@ describe(`nodes db tests`, () => {
     )
     expect(getNodes()).toHaveLength(1)
   })
+
   it(`deletes previously transformed children nodes when parent nodes are deleted`, () => {
     store.dispatch(
       actions.createNode(
@@ -277,39 +282,43 @@ describe(`nodes db tests`, () => {
     )
     expect(getNodes()).toHaveLength(0)
   })
+
   it(`allows deleting nodes`, () => {
-    actions.createNode(
-      {
-        id: `hi`,
-        children: [],
-        parent: `test`,
-        internal: {
-          contentDigest: `hasdfljds`,
-          type: `Test`,
+    store.dispatch(
+      actions.createNode(
+        {
+          id: `hi`,
+          children: [],
+          parent: `test`,
+          internal: {
+            contentDigest: `hasdfljds`,
+            type: `Test`,
+          },
+          pickle: true,
+          deep: {
+            array: [
+              0,
+              1,
+              {
+                boom: true,
+              },
+            ],
+          },
         },
-        pickle: true,
-        deep: {
-          array: [
-            0,
-            1,
-            {
-              boom: true,
-            },
-          ],
-        },
-      },
-      {
-        name: `tests`,
-      }
+        {
+          name: `tests`,
+        }
+      )
     )
-    actions.deleteNode({
-      node: getNode(`hi`),
-    })
+    store.dispatch(
+      actions.deleteNode({
+        node: getNode(`hi`),
+      })
+    )
     expect(getNode(`hi`)).toBeUndefined()
   })
 
   it(`warns when using old deleteNode signature `, () => {
-    console.warn = jest.fn()
     store.dispatch(
       actions.createNode(
         {
@@ -333,9 +342,37 @@ describe(`nodes db tests`, () => {
       })
     )
     expect(getNode(`hi`)).toBeUndefined()
-    const deprecationNotice = `Calling "deleteNode" with a nodeId is deprecated. Please pass an object containing a full node instead: deleteNode({ node })`
-    expect(console.warn).toHaveBeenCalledWith(deprecationNotice)
-    console.warn.mockRestore()
+    const deprecationNotice =
+      `Calling "deleteNode" with a nodeId is deprecated. Please pass an ` +
+      `object containing a full node instead: deleteNode({ node }). ` +
+      `"deleteNode" was called by tests`
+    expect(report.warn).toHaveBeenCalledWith(deprecationNotice)
+  })
+
+  it(`throws an error when trying to delete a node of a type owned from another plugin`, () => {
+    expect(() => {
+      store.dispatch(
+        actions.createNode(
+          {
+            id: `hi`,
+            children: [],
+            parent: `test`,
+            internal: {
+              contentDigest: `hasdfljds`,
+              type: `Other`,
+            },
+          },
+          {
+            name: `other`,
+          }
+        )
+      )
+      store.dispatch(
+        actions.deleteNode(`hi`, getNode(`hi`), {
+          name: `tests`,
+        })
+      )
+    }).toThrow(/deleted/)
   })
 
   it(`does not crash when delete node is called on undefined`, () => {