+++ packages/gatsby/src/utils/babel-parse-to-ast.js
@@ -44,13 +44,13 @@ const PARSER_OPTIONS = {
 }
 
 export function getBabelParserOptions(filePath: string) {
-  // Flow and typescript plugins can't be enabled simultaneously
+  // Flow and TypeScript plugins can't be enabled simultaneously
   if (/\.tsx?/.test(filePath)) {
     const { plugins } = PARSER_OPTIONS
     return {
       ...PARSER_OPTIONS,
-      plugins: plugins.map(
-        plugin => (plugin === `flow` ? `typescript` : plugin)
+      plugins: plugins.map(plugin =>
+        plugin === `flow` ? `typescript` : plugin
       ),
     }
   }