+++ www/src/components/sidebar/item.js
@@ -21,11 +21,12 @@ class Item extends React.PureComponent {
       isActive,
       openSectionHash,
       item,
-      level,
       location,
       onLinkClick,
       onSectionTitleClick,
       ui,
+      isSingle,
+      disableAccordions,
     } = this.props
 
     const isParentOfActiveItem = isItemActive(activeItemParents, item)
@@ -45,25 +46,22 @@ class Item extends React.PureComponent {
             }
             isParentOfActiveItem={isParentOfActiveItem}
             item={item}
-            level={level}
             location={location}
             onLinkClick={onLinkClick}
             openSectionHash={openSectionHash}
             onSectionTitleClick={onSectionTitleClick}
+            isSingle={isSingle}
+            disableAccordions={disableAccordions}
           />
         ) : (
-          <li
-            css={{
-              ...this.props.styles,
-              paddingLeft: level === 0 ? 40 : false,
-            }}
-          >
+          <li>
             {createLink({
               isActive: item.link === activeItemLink.link,
               item,
               location,
               onLinkClick,
-              stepsUI: ui === `steps`,
+              ui,
+              level: item.level,
             })}
           </li>
         )}