+++ packages/gatsby-plugin-react-helmet/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.12](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.11...gatsby-plugin-react-helmet@3.0.12) (2019-04-04)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.0.11](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.10...gatsby-plugin-react-helmet@3.0.11) (2019-03-25)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.0.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.9...gatsby-plugin-react-helmet@3.0.10) (2019-03-16)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.0.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.8...gatsby-plugin-react-helmet@3.0.9) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.7...gatsby-plugin-react-helmet@3.0.8) (2019-03-05)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.0.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.6...gatsby-plugin-react-helmet@3.0.7) (2019-02-25)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+## [3.0.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.5...gatsby-plugin-react-helmet@3.0.6) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
 <a name="3.0.5"></a>
 
 ## [3.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.4...gatsby-plugin-react-helmet@3.0.5) (2018-12-21)